fix media screen
