+++ ios/Sources/MediaFeature/MediaFeature.swift
@@ -1,9 +1,9 @@
 import Component
 import ComposableArchitecture
 import Model
+import Repository
 
 public enum MediaState: Equatable {
-
     case needToInitialize
     case initialized(MediaListState)
 
@@ -12,80 +12,19 @@ public enum MediaState: Equatable {
     }
 }
 
-public struct MediaListState: Equatable {
-
-    enum Next: Equatable {
-        case searchText(String)
-        case isEditingDidChange(Bool)
-        case more(for: MediaType)
-    }
-
-    var blogs: [FeedContent]
-    var videos: [FeedContent]
-    var podcasts: [FeedContent]
-    var next: Next?
-}
-
-public enum MediaType {
-    case blog
-    case video
-    case podcast
-}
-
-public enum MediaAction: Equatable {
-    case loadItems
-    case itemsLoaded([FeedContent])
+public enum MediaAction {
+    case refresh
+    case refreshResponse(Result<[FeedContent], KotlinError>)
     case mediaList(MediaListAction)
 }
 
-public enum MediaListAction: Equatable {
-    case searchTextDidChange(to: String?)
-    case isEditingDidChange(to: Bool)
-    case showMore(for: MediaType)
-    case moreDismissed
-    case tap(FeedContent)
-    case tapFavorite(isFavorited: Bool, id: String)
-}
-
 public struct MediaEnvironment {
-    public init() {}
-}
+    public let feedRepository: FeedRepositoryProtocol
 
-let mediaListReducer = Reducer<MediaListState, MediaListAction, Void> { state, action, _ in
-    switch action {
-    case let .searchTextDidChange(to: searchText):
-        switch state.next {
-        case nil, .searchText, .isEditingDidChange:
-            state.next = searchText.map { .searchText($0) }
-        default:
-            break
-        }
-        return .none
-    case let .isEditingDidChange(isEditing):
-        switch state.next {
-        case nil, .searchText, .isEditingDidChange:
-            state.next = .isEditingDidChange(isEditing)
-            if !isEditing {
-                state.next = nil
-            }
-        default:
-            break
-        }
-        return .none
-    case let .showMore(mediaType):
-        if state.next == nil {
-            state.next = .more(for: mediaType)
-        }
-        return .none
-    case .moreDismissed:
-        if case .more = state.next {
-            state.next = nil
-        }
-        return .none
-    case .tap(let content):
-        return .none
-    case .tapFavorite(let isFavorited, let contentId):
-        return .none
+    public init(
+        feedRepository: FeedRepositoryProtocol
+    ) {
+        self.feedRepository = feedRepository
     }
 }
 
@@ -95,27 +34,26 @@ public let mediaReducer = Reducer<MediaState, MediaAction, MediaEnvironment>.com
         action: /MediaAction.mediaList,
         environment: { _ in () }
     ),
-    .init { state, action, _ in
+    .init { state, action, environment in
         switch action {
-        case .loadItems:
-            // TODO: Load items from the repository
-            return Effect(value: .mockItemsLoads)
-                .delay(for: 1, scheduler: DispatchQueue.main)
-                .eraseToEffect()
-        case let .itemsLoaded(contents):
-            var blogs: [FeedContent] = .init()
-            var videos: [FeedContent] = .init()
-            var podcasts: [FeedContent] = .init()
-            for content in contents {
-                switch content.item.wrappedValue {
+        case .refresh:
+            return environment.feedRepository.feedContents()
+                .catchToEffect()
+                .map(MediaAction.refreshResponse)
+        case let .refreshResponse(.success(feedContents)):
+            var blogs: [FeedContent] = []
+            var videos: [FeedContent] = []
+            var podcasts: [FeedContent] = []
+            for feedContent in feedContents {
+                switch feedContent.item.wrappedValue {
                 case is Blog:
-                    blogs.append(content)
+                    blogs.append(feedContent)
                 case is Video:
-                    videos.append(content)
+                    videos.append(feedContent)
                 case is Podcast:
-                    podcasts.append(content)
+                    podcasts.append(feedContent)
                 default:
-                    assertionFailure("Unexpected FeedItem: (\(content.item.wrappedValue)")
+                    assertionFailure("Unexpected FeedItem: (\(feedContent.item.wrappedValue)")
                     break
                 }
             }
@@ -128,12 +66,12 @@ public let mediaReducer = Reducer<MediaState, MediaAction, MediaEnvironment>.com
                 state = .initialized(.init(blogs: blogs, videos: videos, podcasts: podcasts, next: nil))
             }
             return .none
+        case let .refreshResponse(.failure(error)):
+            print(error.localizedDescription)
+            // TODO: Error handling
+            return .none
         case .mediaList:
             return .none
         }
     }
 )
-
-private extension MediaAction {
-    static let mockItemsLoads: Self = .itemsLoaded([])
-}
